refucturing get handler
