+++ function/src/task/delete.py
@@ -1,16 +1,6 @@
 import logging
-import json
 from task import Task, TaskNotFoundError, NotTaskOwnerError
-from util import fetch_user_id_from_event, convert_return_object, UserNotFoundError, InvalidTaskIdError
-
-
-def fetch_task_id(event):
-    task_id = ''
-    if 'pathParameters' in event and 'task_id' in event['pathParameters']:
-        task_id = event['pathParameters']['task_id']
-        return task_id
-    else:
-        InvalidTaskIdError
+from util import fetch_user_id_from_event, convert_return_object, InvalidTaskIdError, fetch_task_id, UserNotFoundError
 
 
 def lambda_handler(event, context):
@@ -22,6 +12,9 @@ def lambda_handler(event, context):
         raw_task = Task.get(user_id, task_id)
         response = raw_task.delete(user_id)
         return convert_return_object(200, response)
+    except UserNotFoundError as e:
+        logging.error(e)
+        return convert_return_object(400, 'user not found error')
     except InvalidTaskIdError as e:
         logging.error(e)
         return convert_return_object(400, 'invalid request')
