+++ function/src/task/util.py
@@ -9,6 +9,15 @@ class InvalidTaskIdError(Exception):
     pass
 
 
+def fetch_task_id(event):
+    task_id = ''
+    if 'pathParameters' in event and 'task_id' in event['pathParameters']:
+        task_id = event['pathParameters']['task_id']
+        return task_id
+    else:
+        InvalidTaskIdError
+
+
 def fetch_user_id_from_event(event):
     if 'requestContext' in event and 'authorizer' in event['requestContext'] and 'claims' in event['requestContext']['authorizer'] and 'cognito:username' in event['requestContext']['authorizer']['claims']:
         return event['requestContext']['authorizer']['claims']['cognito:username']
